backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pydantic[email]==2.11.1
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.11
bcrypt==4.3.0
sqlalchemy>=1.4.42,<1.5.0
databases[sqlite]==0.7.0
websockets==15.0.1
python-dotenv==1.1.0
aiosqlite==0.21.0
email-validator==2.2.0
